third_party/blink/web_tests/external/wpt/editing/run/multitest_2001-3000-expected.txt

This is a testharness.js-based test.
Found 209 FAIL, 0 TIMEOUT, 0 NOTRUN.
[FAIL] [["subscript",""],["formatblock","<div>"],["inserttext","a"]] "foo[]bar" queryCommandState("subscript") after
  assert_equals: Wrong result returned expected true but got false
[FAIL] [["subscript",""],["formatblock","<div>"],["inserttext","a"]] "foo[]bar" queryCommandValue("subscript") after
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["subscript",""],["formatblock","<div>"],["inserttext","a"]] "foo[]bar" queryCommandValue("formatblock") after
  assert_equals: Wrong result returned expected "div" but got ""
[FAIL] [["subscript",""],["forwarddelete",""]] "foo[]bar" queryCommandValue("subscript") before
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["subscript",""],["forwarddelete",""]] "foo[]bar" queryCommandState("subscript") after
  assert_equals: Wrong result returned expected true but got false
[FAIL] [["subscript",""],["forwarddelete",""]] "foo[]bar" queryCommandValue("subscript") after
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["subscript",""],["forwarddelete",""],["inserttext","a"]] "foo[]bar" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "foo<sub><font size=\\"3\\">a</font></sub>ar" but got "fooaar"
[FAIL] [["subscript",""],["forwarddelete",""],["inserttext","a"]] "foo[]bar" queryCommandValue("subscript") before
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["subscript",""],["forwarddelete",""],["inserttext","a"]] "foo[]bar" queryCommandState("subscript") after
  assert_equals: Wrong result returned expected true but got false
[FAIL] [["subscript",""],["forwarddelete",""],["inserttext","a"]] "foo[]bar" queryCommandValue("subscript") after
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["subscript",""],["indent",""]] "foo[]bar" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<blockquote>foobar</blockquote>" but got "<blockquote style=\\"margin:0 0 0 40px; border:none; padding:0px\\">foobar</blockquote>"
[FAIL] [["subscript",""],["indent",""]] "foo[]bar" queryCommandValue("subscript") before
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["subscript",""],["indent",""]] "foo[]bar" queryCommandState("subscript") after
  assert_equals: Wrong result returned expected true but got false
[FAIL] [["subscript",""],["indent",""]] "foo[]bar" queryCommandValue("subscript") after
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["subscript",""],["indent",""],["inserttext","a"]] "foo[]bar" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<blockquote>foo<sub>a</sub>bar</blockquote>" but got "<blockquote style=\\"margin:0 0 0 40px; border:none; padding:0px\\">fooabar</blockquote>"
[FAIL] [["subscript",""],["indent",""],["inserttext","a"]] "foo[]bar" queryCommandValue("subscript") before
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["subscript",""],["indent",""],["inserttext","a"]] "foo[]bar" queryCommandState("subscript") after
  assert_equals: Wrong result returned expected true but got false
[FAIL] [["subscript",""],["indent",""],["inserttext","a"]] "foo[]bar" queryCommandValue("subscript") after
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["subscript",""],["inserthorizontalrule",""]] "foo[]bar" queryCommandValue("subscript") before
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["subscript",""],["inserthorizontalrule",""]] "foo[]bar" queryCommandState("subscript") after
  assert_equals: Wrong result returned expected true but got false
[FAIL] [["subscript",""],["inserthorizontalrule",""]] "foo[]bar" queryCommandValue("subscript") after
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["subscript",""],["inserthorizontalrule",""],["inserttext","a"]] "foo[]bar" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "foo<hr><sub>a</sub>bar" but got "foo<hr>abar"
[FAIL] [["subscript",""],["inserthorizontalrule",""],["inserttext","a"]] "foo[]bar" queryCommandValue("subscript") before
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["subscript",""],["inserthorizontalrule",""],["inserttext","a"]] "foo[]bar" queryCommandState("subscript") after
  assert_equals: Wrong result returned expected true but got false
[FAIL] [["subscript",""],["inserthorizontalrule",""],["inserttext","a"]] "foo[]bar" queryCommandValue("subscript") after
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["subscript",""],["inserthtml","ab<b>c</b>d"]] "foo[]bar" queryCommandValue("subscript") before
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["subscript",""],["inserthtml","ab<b>c</b>d"]] "foo[]bar" queryCommandState("subscript") after
  assert_equals: Wrong result returned expected true but got false
[FAIL] [["subscript",""],["inserthtml","ab<b>c</b>d"]] "foo[]bar" queryCommandValue("subscript") after
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["subscript",""],["inserthtml","ab<b>c</b>d"],["inserttext","a"]] "foo[]bar" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "fooab<b>c</b>d<sub>a</sub>bar" but got "fooab<b>c</b>dabar"
[FAIL] [["subscript",""],["inserthtml","ab<b>c</b>d"],["inserttext","a"]] "foo[]bar" queryCommandValue("subscript") before
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["subscript",""],["inserthtml","ab<b>c</b>d"],["inserttext","a"]] "foo[]bar" queryCommandState("subscript") after
  assert_equals: Wrong result returned expected true but got false
[FAIL] [["subscript",""],["inserthtml","ab<b>c</b>d"],["inserttext","a"]] "foo[]bar" queryCommandValue("subscript") after
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["subscript",""],["insertimage","/img/lion.svg"]] "foo[]bar" queryCommandValue("subscript") before
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["subscript",""],["insertimage","/img/lion.svg"]] "foo[]bar" queryCommandState("subscript") after
  assert_equals: Wrong result returned expected true but got false
[FAIL] [["subscript",""],["insertimage","/img/lion.svg"]] "foo[]bar" queryCommandValue("subscript") after
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["subscript",""],["insertimage","/img/lion.svg"],["inserttext","a"]] "foo[]bar" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "foo<img src=\\"/img/lion.svg\\"><sub>a</sub>bar" but got "foo<img src=\\"/img/lion.svg\\">abar"
[FAIL] [["subscript",""],["insertimage","/img/lion.svg"],["inserttext","a"]] "foo[]bar" queryCommandValue("subscript") before
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["subscript",""],["insertimage","/img/lion.svg"],["inserttext","a"]] "foo[]bar" queryCommandState("subscript") after
  assert_equals: Wrong result returned expected true but got false
[FAIL] [["subscript",""],["insertimage","/img/lion.svg"],["inserttext","a"]] "foo[]bar" queryCommandValue("subscript") after
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["subscript",""],["insertlinebreak",""]] "foo[]bar" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "foo<br>bar" but got "foo<span style=\\"vertical-align:sub\\"><br></span>bar"
[FAIL] [["subscript",""],["insertlinebreak",""]] "foo[]bar" queryCommandValue("subscript") before
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["subscript",""],["insertlinebreak",""]] "foo[]bar" queryCommandValue("subscript") after
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["subscript",""],["insertlinebreak",""],["inserttext","a"]] "foo[]bar" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "foo<br><sub>a</sub>bar" but got "foo<span style=\\"vertical-align:sub\\"><br>a</span>bar"
[FAIL] [["subscript",""],["insertlinebreak",""],["inserttext","a"]] "foo[]bar" queryCommandValue("subscript") before
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["subscript",""],["insertlinebreak",""],["inserttext","a"]] "foo[]bar" queryCommandValue("subscript") after
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["subscript",""],["insertorderedlist",""]] "foo[]bar" queryCommandValue("subscript") before
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["subscript",""],["insertorderedlist",""]] "foo[]bar" queryCommandState("subscript") after
  assert_equals: Wrong result returned expected true but got false
[FAIL] [["subscript",""],["insertorderedlist",""]] "foo[]bar" queryCommandValue("subscript") after
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["subscript",""],["insertorderedlist",""]] "foo[]bar" queryCommandValue("insertorderedlist") before
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["subscript",""],["insertorderedlist",""]] "foo[]bar" queryCommandValue("insertorderedlist") after
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["subscript",""],["insertorderedlist",""],["inserttext","a"]] "foo[]bar" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<ol><li>foo<sub>a</sub>bar</li></ol>" but got "<ol><li>afoobar</li></ol>"
[FAIL] [["subscript",""],["insertorderedlist",""],["inserttext","a"]] "foo[]bar" queryCommandValue("subscript") before
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["subscript",""],["insertorderedlist",""],["inserttext","a"]] "foo[]bar" queryCommandState("subscript") after
  assert_equals: Wrong result returned expected true but got false
[FAIL] [["subscript",""],["insertorderedlist",""],["inserttext","a"]] "foo[]bar" queryCommandValue("subscript") after
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["subscript",""],["insertorderedlist",""],["inserttext","a"]] "foo[]bar" queryCommandValue("insertorderedlist") before
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["subscript",""],["insertorderedlist",""],["inserttext","a"]] "foo[]bar" queryCommandValue("insertorderedlist") after
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["subscript",""],["insertparagraph",""]] "foo[]bar" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<p>foo</p><p>bar</p>" but got "foo<div>bar</div>"
[FAIL] [["subscript",""],["insertparagraph",""]] "foo[]bar" queryCommandValue("subscript") before
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["subscript",""],["insertparagraph",""]] "foo[]bar" queryCommandValue("subscript") after
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["subscript",""],["insertparagraph",""],["inserttext","a"]] "foo[]bar" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<p>foo</p><p><sub>a</sub>bar</p>" but got "foo<div><span style=\\"vertical-align:sub\\">a</span>bar</div>"
[FAIL] [["subscript",""],["insertparagraph",""],["inserttext","a"]] "foo[]bar" queryCommandValue("subscript") before
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["subscript",""],["insertparagraph",""],["inserttext","a"]] "foo[]bar" queryCommandValue("subscript") after
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["subscript",""],["insertunorderedlist",""]] "foo[]bar" queryCommandValue("subscript") before
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["subscript",""],["insertunorderedlist",""]] "foo[]bar" queryCommandState("subscript") after
  assert_equals: Wrong result returned expected true but got false
[FAIL] [["subscript",""],["insertunorderedlist",""]] "foo[]bar" queryCommandValue("subscript") after
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["subscript",""],["insertunorderedlist",""]] "foo[]bar" queryCommandValue("insertunorderedlist") before
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["subscript",""],["insertunorderedlist",""]] "foo[]bar" queryCommandValue("insertunorderedlist") after
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["subscript",""],["insertunorderedlist",""],["inserttext","a"]] "foo[]bar" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<ul><li>foo<sub>a</sub>bar</li></ul>" but got "<ul><li>afoobar</li></ul>"
[FAIL] [["subscript",""],["insertunorderedlist",""],["inserttext","a"]] "foo[]bar" queryCommandValue("subscript") before
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["subscript",""],["insertunorderedlist",""],["inserttext","a"]] "foo[]bar" queryCommandState("subscript") after
  assert_equals: Wrong result returned expected true but got false
[FAIL] [["subscript",""],["insertunorderedlist",""],["inserttext","a"]] "foo[]bar" queryCommandValue("subscript") after
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["subscript",""],["insertunorderedlist",""],["inserttext","a"]] "foo[]bar" queryCommandValue("insertunorderedlist") before
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["subscript",""],["insertunorderedlist",""],["inserttext","a"]] "foo[]bar" queryCommandValue("insertunorderedlist") after
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["subscript",""],["justifycenter",""]] "foo[]bar" queryCommandValue("subscript") before
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["subscript",""],["justifycenter",""]] "foo[]bar" queryCommandState("subscript") after
  assert_equals: Wrong result returned expected true but got false
[FAIL] [["subscript",""],["justifycenter",""]] "foo[]bar" queryCommandValue("subscript") after
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["subscript",""],["justifycenter",""]] "foo[]bar" queryCommandValue("justifycenter") before
  assert_equals: Wrong result returned expected "left" but got "false"
[FAIL] [["subscript",""],["justifycenter",""]] "foo[]bar" queryCommandValue("justifycenter") after
  assert_equals: Wrong result returned expected "center" but got "true"
[FAIL] [["subscript",""],["justifycenter",""],["inserttext","a"]] "foo[]bar" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<div style=\\"text-align:center\\">foo<sub>a</sub>bar</div>" but got "<div style=\\"text-align:center\\">fooabar</div>"
[FAIL] [["subscript",""],["justifycenter",""],["inserttext","a"]] "foo[]bar" queryCommandValue("subscript") before
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["subscript",""],["justifycenter",""],["inserttext","a"]] "foo[]bar" queryCommandState("subscript") after
  assert_equals: Wrong result returned expected true but got false
[FAIL] [["subscript",""],["justifycenter",""],["inserttext","a"]] "foo[]bar" queryCommandValue("subscript") after
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["subscript",""],["justifycenter",""],["inserttext","a"]] "foo[]bar" queryCommandValue("justifycenter") before
  assert_equals: Wrong result returned expected "left" but got "false"
[FAIL] [["subscript",""],["justifycenter",""],["inserttext","a"]] "foo[]bar" queryCommandValue("justifycenter") after
  assert_equals: Wrong result returned expected "center" but got "true"
[FAIL] [["subscript",""],["justifyfull",""]] "foo[]bar" queryCommandValue("subscript") before
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["subscript",""],["justifyfull",""]] "foo[]bar" queryCommandState("subscript") after
  assert_equals: Wrong result returned expected true but got false
[FAIL] [["subscript",""],["justifyfull",""]] "foo[]bar" queryCommandValue("subscript") after
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["subscript",""],["justifyfull",""]] "foo[]bar" queryCommandValue("justifyfull") before
  assert_equals: Wrong result returned expected "left" but got "false"
[FAIL] [["subscript",""],["justifyfull",""]] "foo[]bar" queryCommandValue("justifyfull") after
  assert_equals: Wrong result returned expected "justify" but got "true"
[FAIL] [["subscript",""],["justifyfull",""],["inserttext","a"]] "foo[]bar" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<div style=\\"text-align:justify\\">foo<sub>a</sub>bar</div>" but got "<div style=\\"text-align:justify\\">fooabar</div>"
[FAIL] [["subscript",""],["justifyfull",""],["inserttext","a"]] "foo[]bar" queryCommandValue("subscript") before
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["subscript",""],["justifyfull",""],["inserttext","a"]] "foo[]bar" queryCommandState("subscript") after
  assert_equals: Wrong result returned expected true but got false
[FAIL] [["subscript",""],["justifyfull",""],["inserttext","a"]] "foo[]bar" queryCommandValue("subscript") after
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["subscript",""],["justifyfull",""],["inserttext","a"]] "foo[]bar" queryCommandValue("justifyfull") before
  assert_equals: Wrong result returned expected "left" but got "false"
[FAIL] [["subscript",""],["justifyfull",""],["inserttext","a"]] "foo[]bar" queryCommandValue("justifyfull") after
  assert_equals: Wrong result returned expected "justify" but got "true"
[FAIL] [["subscript",""],["justifyleft",""]] "foo[]bar" queryCommandValue("subscript") before
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["subscript",""],["justifyleft",""]] "foo[]bar" queryCommandState("subscript") after
  assert_equals: Wrong result returned expected true but got false
[FAIL] [["subscript",""],["justifyleft",""]] "foo[]bar" queryCommandValue("subscript") after
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["subscript",""],["justifyleft",""]] "foo[]bar" queryCommandValue("justifyleft") before
  assert_equals: Wrong result returned expected "left" but got "true"
[FAIL] [["subscript",""],["justifyleft",""]] "foo[]bar" queryCommandValue("justifyleft") after
  assert_equals: Wrong result returned expected "left" but got "true"
[FAIL] [["subscript",""],["justifyleft",""],["inserttext","a"]] "foo[]bar" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "foo<sub>a</sub>bar" but got "fooabar"
[FAIL] [["subscript",""],["justifyleft",""],["inserttext","a"]] "foo[]bar" queryCommandValue("subscript") before
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["subscript",""],["justifyleft",""],["inserttext","a"]] "foo[]bar" queryCommandState("subscript") after
  assert_equals: Wrong result returned expected true but got false
[FAIL] [["subscript",""],["justifyleft",""],["inserttext","a"]] "foo[]bar" queryCommandValue("subscript") after
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["subscript",""],["justifyleft",""],["inserttext","a"]] "foo[]bar" queryCommandValue("justifyleft") before
  assert_equals: Wrong result returned expected "left" but got "true"
[FAIL] [["subscript",""],["justifyleft",""],["inserttext","a"]] "foo[]bar" queryCommandValue("justifyleft") after
  assert_equals: Wrong result returned expected "left" but got "true"
[FAIL] [["subscript",""],["justifyright",""]] "foo[]bar" queryCommandValue("subscript") before
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["subscript",""],["justifyright",""]] "foo[]bar" queryCommandState("subscript") after
  assert_equals: Wrong result returned expected true but got false
[FAIL] [["subscript",""],["justifyright",""]] "foo[]bar" queryCommandValue("subscript") after
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["subscript",""],["justifyright",""]] "foo[]bar" queryCommandValue("justifyright") before
  assert_equals: Wrong result returned expected "left" but got "false"
[FAIL] [["subscript",""],["justifyright",""]] "foo[]bar" queryCommandValue("justifyright") after
  assert_equals: Wrong result returned expected "right" but got "true"
[FAIL] [["subscript",""],["justifyright",""],["inserttext","a"]] "foo[]bar" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<div style=\\"text-align:right\\">foo<sub>a</sub>bar</div>" but got "<div style=\\"text-align:right\\">fooabar</div>"
[FAIL] [["subscript",""],["justifyright",""],["inserttext","a"]] "foo[]bar" queryCommandValue("subscript") before
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["subscript",""],["justifyright",""],["inserttext","a"]] "foo[]bar" queryCommandState("subscript") after
  assert_equals: Wrong result returned expected true but got false
[FAIL] [["subscript",""],["justifyright",""],["inserttext","a"]] "foo[]bar" queryCommandValue("subscript") after
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["subscript",""],["justifyright",""],["inserttext","a"]] "foo[]bar" queryCommandValue("justifyright") before
  assert_equals: Wrong result returned expected "left" but got "false"
[FAIL] [["subscript",""],["justifyright",""],["inserttext","a"]] "foo[]bar" queryCommandValue("justifyright") after
  assert_equals: Wrong result returned expected "right" but got "true"
[FAIL] [["subscript",""],["outdent",""]] "foo[]bar" queryCommandValue("subscript") before
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["subscript",""],["outdent",""]] "foo[]bar" queryCommandValue("subscript") after
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["subscript",""],["outdent",""],["inserttext","a"]] "foo[]bar" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "foo<sub>a</sub>bar" but got "foo<span style=\\"vertical-align:sub\\">a</span>bar"
[FAIL] [["subscript",""],["outdent",""],["inserttext","a"]] "foo[]bar" queryCommandValue("subscript") before
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["subscript",""],["outdent",""],["inserttext","a"]] "foo[]bar" queryCommandValue("subscript") after
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["superscript",""],["inserttext","a"]] "foo[]bar" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "foo<sup>a</sup>bar" but got "foo<span style=\\"vertical-align:super\\">a</span>bar"
[FAIL] [["superscript",""],["inserttext","a"]] "foo[]bar" queryCommandValue("superscript") before
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["superscript",""],["inserttext","a"]] "foo[]bar" queryCommandValue("superscript") after
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["superscript",""],["delete",""]] "foo[]bar" queryCommandValue("superscript") before
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["superscript",""],["delete",""]] "foo[]bar" queryCommandState("superscript") after
  assert_equals: Wrong result returned expected true but got false
[FAIL] [["superscript",""],["delete",""]] "foo[]bar" queryCommandValue("superscript") after
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["superscript",""],["delete",""],["inserttext","a"]] "foo[]bar" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "fo<sup><font size=\\"3\\">a</font></sup>bar" but got "foabar"
[FAIL] [["superscript",""],["delete",""],["inserttext","a"]] "foo[]bar" queryCommandValue("superscript") before
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["superscript",""],["delete",""],["inserttext","a"]] "foo[]bar" queryCommandState("superscript") after
  assert_equals: Wrong result returned expected true but got false
[FAIL] [["superscript",""],["delete",""],["inserttext","a"]] "foo[]bar" queryCommandValue("superscript") after
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["superscript",""],["formatblock","<div>"]] "foo[]bar" queryCommandValue("superscript") before
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["superscript",""],["formatblock","<div>"]] "foo[]bar" queryCommandState("superscript") after
  assert_equals: Wrong result returned expected true but got false
[FAIL] [["superscript",""],["formatblock","<div>"]] "foo[]bar" queryCommandValue("superscript") after
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["superscript",""],["formatblock","<div>"],["inserttext","a"]] "foo[]bar" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<div>foo<sup>a</sup>bar</div>" but got "<div>fooabar</div>"
[FAIL] [["superscript",""],["formatblock","<div>"],["inserttext","a"]] "foo[]bar" queryCommandValue("superscript") before
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["superscript",""],["formatblock","<div>"],["inserttext","a"]] "foo[]bar" queryCommandState("superscript") after
  assert_equals: Wrong result returned expected true but got false
[FAIL] [["superscript",""],["formatblock","<div>"],["inserttext","a"]] "foo[]bar" queryCommandValue("superscript") after
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["superscript",""],["forwarddelete",""]] "foo[]bar" queryCommandValue("superscript") before
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["superscript",""],["forwarddelete",""]] "foo[]bar" queryCommandState("superscript") after
  assert_equals: Wrong result returned expected true but got false
[FAIL] [["superscript",""],["forwarddelete",""]] "foo[]bar" queryCommandValue("superscript") after
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["superscript",""],["forwarddelete",""],["inserttext","a"]] "foo[]bar" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "foo<sup><font size=\\"3\\">a</font></sup>ar" but got "fooaar"
[FAIL] [["superscript",""],["forwarddelete",""],["inserttext","a"]] "foo[]bar" queryCommandValue("superscript") before
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["superscript",""],["forwarddelete",""],["inserttext","a"]] "foo[]bar" queryCommandState("superscript") after
  assert_equals: Wrong result returned expected true but got false
[FAIL] [["superscript",""],["forwarddelete",""],["inserttext","a"]] "foo[]bar" queryCommandValue("superscript") after
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["superscript",""],["indent",""]] "foo[]bar" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<blockquote>foobar</blockquote>" but got "<blockquote style=\\"margin:0 0 0 40px; border:none; padding:0px\\">foobar</blockquote>"
[FAIL] [["superscript",""],["indent",""]] "foo[]bar" queryCommandValue("superscript") before
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["superscript",""],["indent",""]] "foo[]bar" queryCommandState("superscript") after
  assert_equals: Wrong result returned expected true but got false
[FAIL] [["superscript",""],["indent",""]] "foo[]bar" queryCommandValue("superscript") after
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["superscript",""],["indent",""],["inserttext","a"]] "foo[]bar" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<blockquote>foo<sup>a</sup>bar</blockquote>" but got "<blockquote style=\\"margin:0 0 0 40px; border:none; padding:0px\\">fooabar</blockquote>"
[FAIL] [["superscript",""],["indent",""],["inserttext","a"]] "foo[]bar" queryCommandValue("superscript") before
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["superscript",""],["indent",""],["inserttext","a"]] "foo[]bar" queryCommandState("superscript") after
  assert_equals: Wrong result returned expected true but got false
[FAIL] [["superscript",""],["indent",""],["inserttext","a"]] "foo[]bar" queryCommandValue("superscript") after
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["superscript",""],["inserthorizontalrule",""]] "foo[]bar" queryCommandValue("superscript") before
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["superscript",""],["inserthorizontalrule",""]] "foo[]bar" queryCommandState("superscript") after
  assert_equals: Wrong result returned expected true but got false
[FAIL] [["superscript",""],["inserthorizontalrule",""]] "foo[]bar" queryCommandValue("superscript") after
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["superscript",""],["inserthorizontalrule",""],["inserttext","a"]] "foo[]bar" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "foo<hr><sup>a</sup>bar" but got "foo<hr>abar"
[FAIL] [["superscript",""],["inserthorizontalrule",""],["inserttext","a"]] "foo[]bar" queryCommandValue("superscript") before
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["superscript",""],["inserthorizontalrule",""],["inserttext","a"]] "foo[]bar" queryCommandState("superscript") after
  assert_equals: Wrong result returned expected true but got false
[FAIL] [["superscript",""],["inserthorizontalrule",""],["inserttext","a"]] "foo[]bar" queryCommandValue("superscript") after
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["superscript",""],["inserthtml","ab<b>c</b>d"]] "foo[]bar" queryCommandValue("superscript") before
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["superscript",""],["inserthtml","ab<b>c</b>d"]] "foo[]bar" queryCommandState("superscript") after
  assert_equals: Wrong result returned expected true but got false
[FAIL] [["superscript",""],["inserthtml","ab<b>c</b>d"]] "foo[]bar" queryCommandValue("superscript") after
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["superscript",""],["inserthtml","ab<b>c</b>d"],["inserttext","a"]] "foo[]bar" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "fooab<b>c</b>d<sup>a</sup>bar" but got "fooab<b>c</b>dabar"
[FAIL] [["superscript",""],["inserthtml","ab<b>c</b>d"],["inserttext","a"]] "foo[]bar" queryCommandValue("superscript") before
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["superscript",""],["inserthtml","ab<b>c</b>d"],["inserttext","a"]] "foo[]bar" queryCommandState("superscript") after
  assert_equals: Wrong result returned expected true but got false
[FAIL] [["superscript",""],["inserthtml","ab<b>c</b>d"],["inserttext","a"]] "foo[]bar" queryCommandValue("superscript") after
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["superscript",""],["insertimage","/img/lion.svg"]] "foo[]bar" queryCommandValue("superscript") before
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["superscript",""],["insertimage","/img/lion.svg"]] "foo[]bar" queryCommandState("superscript") after
  assert_equals: Wrong result returned expected true but got false
[FAIL] [["superscript",""],["insertimage","/img/lion.svg"]] "foo[]bar" queryCommandValue("superscript") after
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["superscript",""],["insertimage","/img/lion.svg"],["inserttext","a"]] "foo[]bar" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "foo<img src=\\"/img/lion.svg\\"><sup>a</sup>bar" but got "foo<img src=\\"/img/lion.svg\\">abar"
[FAIL] [["superscript",""],["insertimage","/img/lion.svg"],["inserttext","a"]] "foo[]bar" queryCommandValue("superscript") before
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["superscript",""],["insertimage","/img/lion.svg"],["inserttext","a"]] "foo[]bar" queryCommandState("superscript") after
  assert_equals: Wrong result returned expected true but got false
[FAIL] [["superscript",""],["insertimage","/img/lion.svg"],["inserttext","a"]] "foo[]bar" queryCommandValue("superscript") after
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["superscript",""],["insertlinebreak",""]] "foo[]bar" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "foo<br>bar" but got "foo<span style=\\"vertical-align:super\\"><br></span>bar"
[FAIL] [["superscript",""],["insertlinebreak",""]] "foo[]bar" queryCommandValue("superscript") before
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["superscript",""],["insertlinebreak",""]] "foo[]bar" queryCommandValue("superscript") after
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["superscript",""],["insertlinebreak",""],["inserttext","a"]] "foo[]bar" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "foo<br><sup>a</sup>bar" but got "foo<span style=\\"vertical-align:super\\"><br>a</span>bar"
[FAIL] [["superscript",""],["insertlinebreak",""],["inserttext","a"]] "foo[]bar" queryCommandValue("superscript") before
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["superscript",""],["insertlinebreak",""],["inserttext","a"]] "foo[]bar" queryCommandValue("superscript") after
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["superscript",""],["insertorderedlist",""]] "foo[]bar" queryCommandValue("superscript") before
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["superscript",""],["insertorderedlist",""]] "foo[]bar" queryCommandState("superscript") after
  assert_equals: Wrong result returned expected true but got false
[FAIL] [["superscript",""],["insertorderedlist",""]] "foo[]bar" queryCommandValue("superscript") after
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["superscript",""],["insertorderedlist",""]] "foo[]bar" queryCommandValue("insertorderedlist") before
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["superscript",""],["insertorderedlist",""]] "foo[]bar" queryCommandValue("insertorderedlist") after
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["superscript",""],["insertorderedlist",""],["inserttext","a"]] "foo[]bar" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<ol><li>foo<sup>a</sup>bar</li></ol>" but got "<ol><li>afoobar</li></ol>"
[FAIL] [["superscript",""],["insertorderedlist",""],["inserttext","a"]] "foo[]bar" queryCommandValue("superscript") before
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["superscript",""],["insertorderedlist",""],["inserttext","a"]] "foo[]bar" queryCommandState("superscript") after
  assert_equals: Wrong result returned expected true but got false
[FAIL] [["superscript",""],["insertorderedlist",""],["inserttext","a"]] "foo[]bar" queryCommandValue("superscript") after
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["superscript",""],["insertorderedlist",""],["inserttext","a"]] "foo[]bar" queryCommandValue("insertorderedlist") before
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["superscript",""],["insertorderedlist",""],["inserttext","a"]] "foo[]bar" queryCommandValue("insertorderedlist") after
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["superscript",""],["insertparagraph",""]] "foo[]bar" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<p>foo</p><p>bar</p>" but got "foo<div>bar</div>"
[FAIL] [["superscript",""],["insertparagraph",""]] "foo[]bar" queryCommandValue("superscript") before
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["superscript",""],["insertparagraph",""]] "foo[]bar" queryCommandValue("superscript") after
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["superscript",""],["insertparagraph",""],["inserttext","a"]] "foo[]bar" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<p>foo</p><p><sup>a</sup>bar</p>" but got "foo<div><span style=\\"vertical-align:super\\">a</span>bar</div>"
[FAIL] [["superscript",""],["insertparagraph",""],["inserttext","a"]] "foo[]bar" queryCommandValue("superscript") before
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["superscript",""],["insertparagraph",""],["inserttext","a"]] "foo[]bar" queryCommandValue("superscript") after
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["superscript",""],["insertunorderedlist",""]] "foo[]bar" queryCommandValue("superscript") before
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["superscript",""],["insertunorderedlist",""]] "foo[]bar" queryCommandState("superscript") after
  assert_equals: Wrong result returned expected true but got false
[FAIL] [["superscript",""],["insertunorderedlist",""]] "foo[]bar" queryCommandValue("superscript") after
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["superscript",""],["insertunorderedlist",""]] "foo[]bar" queryCommandValue("insertunorderedlist") before
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["superscript",""],["insertunorderedlist",""]] "foo[]bar" queryCommandValue("insertunorderedlist") after
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["superscript",""],["insertunorderedlist",""],["inserttext","a"]] "foo[]bar" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<ul><li>foo<sup>a</sup>bar</li></ul>" but got "<ul><li>afoobar</li></ul>"
[FAIL] [["superscript",""],["insertunorderedlist",""],["inserttext","a"]] "foo[]bar" queryCommandValue("superscript") before
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["superscript",""],["insertunorderedlist",""],["inserttext","a"]] "foo[]bar" queryCommandState("superscript") after
  assert_equals: Wrong result returned expected true but got false
[FAIL] [["superscript",""],["insertunorderedlist",""],["inserttext","a"]] "foo[]bar" queryCommandValue("superscript") after
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["superscript",""],["insertunorderedlist",""],["inserttext","a"]] "foo[]bar" queryCommandValue("insertunorderedlist") before
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["superscript",""],["insertunorderedlist",""],["inserttext","a"]] "foo[]bar" queryCommandValue("insertunorderedlist") after
  assert_equals: Wrong result returned expected "" but got "true"
Harness: the test ran to completion.
